requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.19.0
pillow>=9.0.0
pyyaml>=6.0
numpy>=1.24.0 